Initialize CVS branch CSRG
The per-file CVS branch roots do not exactly match a reconstructed source-branch commit.

+++ sys/sys/ioccom.h
@@ -1,5 +1,3 @@
-/*	$NetBSD: ioccom.h,v 1.4 1994/10/30 21:49:56 cgd Exp $	*/
-
 /*-
  * Copyright (c) 1982, 1986, 1990, 1993, 1994
  *	The Regents of the University of California.  All rights reserved.
@@ -32,7 +30,8 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)ioccom.h	8.2 (Berkeley) 3/28/94
+ *	from: @(#)ioccom.h	8.2 (Berkeley) 3/28/94
+ *	$Id: ioccom.h,v 1.1 1994/05/21 03:51:59 cgd Exp $
  */
 
 #ifndef	_SYS_IOCCOM_H_
@@ -48,17 +47,12 @@
 #define	IOCBASECMD(x)	((x) & ~(IOCPARM_MASK << 16))
 #define	IOCGROUP(x)	(((x) >> 8) & 0xff)
 
-#define	IOCPARM_MAX	NBPG	/* max size of ioctl args, mult. of NBPG */
-				/* no parameters */
-#define	IOC_VOID	(unsigned long)0x20000000
-				/* copy parameters out */
-#define	IOC_OUT		(unsigned long)0x40000000
-				/* copy parameters in */
-#define	IOC_IN		(unsigned long)0x80000000
-				/* copy paramters in and out */
+#define	IOCPARM_MAX	NBPG		/* max size of ioctl, mult. of NBPG */
+#define	IOC_VOID	0x20000000	/* no parameters */
+#define	IOC_OUT		0x40000000	/* copy out parameters */
+#define	IOC_IN		0x80000000	/* copy in parameters */
 #define	IOC_INOUT	(IOC_IN|IOC_OUT)
-				/* mask for IN/OUT/VOID */
-#define	IOC_DIRMASK	(unsigned long)0xe0000000
+#define	IOC_DIRMASK	0xe0000000	/* mask for IN/OUT/VOID */
 
 #define	_IOC(inout,group,num,len) \
 	(inout | ((len & IOCPARM_MASK) << 16) | ((group) << 8) | (num))
